backend/pyproject.toml
[build-system]
requires = ["setuptools>=67.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "quizpoolai-backend"
version = "0.1.0"
description = "FastAPI backend for the YouTube → Quiz generator."
readme = "../README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
dependencies = [
    "fastapi==0.111.0",
    "uvicorn[standard]==0.30.1",
    "youtube-transcript-api==0.6.2",
    "google-generativeai==0.5.2",
    "pinecone-client==4.1.1",
    "yt-dlp==2024.4.9",
    "openai-whisper==20231117",
    "python-dotenv==1.0.1",
    "pydantic==1.10.15",
    "pytest==8.2.2",
    "httpx==0.27.0",
    "ruff==0.4.6",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "B", "I"]
